+++ changelog.md
@@ -3,6 +3,7 @@
   * The `insanesurvivaloverhaul:creative_removal` item tag can be used to remove items from the creative inventory
 * Undead now burn in daylight even if there's a block above them
 * Enchanted items can no longer be used for anvil crafting
+* Fixed Sunlight modifier not taking into account the day time (it only checked for sky light)
 
 ## 5.4.3.1
 * Sprinting consumes much more hunger (5x)